proj3 big seed problems fixed

+++ proj3/byow/Core/Engine.java
@@ -44,7 +44,7 @@ public class Engine {
         // See proj3.byow.InputDemo for a demo of how you can make a nice clean interface
         // that works for many different input types.
 
-        int seed = Integer.parseInt(input.substring(1, input.length() - 1));
+        long seed = Integer.parseInt(input.substring(1, input.length() - 1));
         random = new Random(seed);
         Floor floor = new Floor(1);
         FloorManager floorManager = new FloorManager(floor);
